chore: move heading logic from tabs transform to util file
